correccion de la segunda entrega
