+++ Entrega2/data/fs/files/products.json
@@ -1,0 +1,11 @@
+[
+  {
+    "title": "aaa"
+  },
+  {
+    "price": 1000
+  },
+  {
+    "stock": 200
+  }
+]
