+++ Entrega2/data/fs/files/user.json
@@ -1,0 +1,11 @@
+[
+  {
+    "name": "aa"
+  },
+  {
+    "lastname": "bbb"
+  },
+  {
+    "gmail": "ffff"
+  }
+]
